requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
